Split ABI mapping and blockchain invocation logic into new files
