+++ src/event-stream/event-stream.service.ts
@@ -266,7 +266,6 @@ export class EventStreamService {
     instancePath: string,
     eventABI: IAbiMethod,
     streamId: string,
-    event: string,
     name: string,
     address: string,
     methods: IAbiMethod[],
@@ -286,7 +285,7 @@ export class EventStreamService {
         this.requestOptions(ctx),
       ),
     );
-    this.logger.log(`Created subscription ${event}: ${response.data.id}`);
+    this.logger.log(`Created subscription ${name}: ${response.data.id}`);
     return response.data;
   }
 
@@ -295,7 +294,6 @@ export class EventStreamService {
     instancePath: string,
     eventABI: IAbiMethod,
     streamId: string,
-    event: string,
     name: string,
     contractAddress: string,
     possibleABIs: IAbiMethod[],
@@ -304,7 +302,7 @@ export class EventStreamService {
     const existingSubscriptions = await this.getSubscriptions(ctx);
     const sub = existingSubscriptions.find(s => s.name === name && s.stream === streamId);
     if (sub) {
-      this.logger.log(`Existing subscription for ${event}: ${sub.id}`);
+      this.logger.log(`Existing subscription for ${name}: ${sub.id}`);
       return sub;
     }
     return this.createSubscription(
@@ -312,7 +310,6 @@ export class EventStreamService {
       instancePath,
       eventABI,
       streamId,
-      event,
       name,
       contractAddress,
       possibleABIs,
