+++ src/health/health.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get } from '@nestjs/common';
 import { HealthCheckService, HealthCheck, HttpHealthIndicator } from '@nestjs/terminus';
-import { TokensService } from '../tokens/tokens.service';
+import { BlockchainConnectorService } from '../tokens/blockchain.service';
 import { basicAuth } from '../utils';
 
 @Controller('health')
@@ -8,7 +8,7 @@ export class HealthController {
   constructor(
     private health: HealthCheckService,
     private http: HttpHealthIndicator,
-    private readonly tokensService: TokensService,
+    private blockchain: BlockchainConnectorService,
   ) {}
 
   @Get('/liveness')
@@ -24,8 +24,8 @@ export class HealthController {
       () =>
         this.http.pingCheck(
           'ethconnect',
-          `${this.tokensService.baseUrl}/status`,
-          basicAuth(this.tokensService.username, this.tokensService.password),
+          `${this.blockchain.baseUrl}/status`,
+          basicAuth(this.blockchain.username, this.blockchain.password),
         ),
     ]);
   }
