+++ test/app.e2e-context.ts
@@ -12,6 +12,7 @@ import { EventStreamService } from '../src/event-stream/event-stream.service';
 import { EventStreamProxyGateway } from '../src/eventstream-proxy/eventstream-proxy.gateway';
 import { TokensService } from '../src/tokens/tokens.service';
 import { requestIDMiddleware } from '../src/request-context/request-id.middleware';
+import { BlockchainConnectorService } from '../src/tokens/blockchain.service';
 
 export const BASE_URL = 'http://eth';
 export const INSTANCE_PATH = '/tokens';
@@ -70,7 +71,8 @@ export class TestContext {
     await this.app.init();
 
     this.app.get(EventStreamProxyGateway).configure('url', TOPIC);
-    this.app.get(TokensService).configure(BASE_URL, '', TOPIC, PREFIX, '', '', '', []);
+    this.app.get(TokensService).configure(BASE_URL, TOPIC, PREFIX, '');
+    this.app.get(BlockchainConnectorService).configure(BASE_URL, '', '', '', []);
 
     (this.app.getHttpServer() as Server).listen();
     this.server = request(this.app.getHttpServer());
